# ops/api/dappswagger.yaml
openapi: 3.0.0
servers:
  - url: 'http://52.15.91.107:3002/api/v1'
info:
  description: >-
    除注册与登陆两个API外，其他API的访问均需要在request header加上JWT token，否则返回`401 Unauthorized`。

    登陆成功后Server会返回JWT token。

    JWT token的格式为：
      - Value `Authorization: Bearer <token>` from request header

    **注意Bearer和token之前有一个空格**


    当使用swagger-ui访问API时，需要点右下方绿色`Authorize`按钮，输入`Bearer <token>`
    
    
    - 铸币账户：新建的token全部在这个账户里
    
    - gas账户：交易产生的所有gas都存到这个账户里
    
    - ram账户：所有ram都从这个账户里支出
    
    
    
    Sample Dapp code: https://github.com/XDMu/SampleDapp
    
    

    # change log __v0.0.2__
    
    本次版本主要更新如下：
    
    - Earth部署成功后需要初始化，初始化之前需要创建两个Earth账户，作为GZH的铸币账户和gas账户
    
    - 只有Admin才能创建token，创建token时需要指定3个账户  
    
    - 交易有gas和ram
    
    - admin可以改ram
    
    - 铸币账户可以改gas
    
    - Dapp部署之后需要到Earth注册一个合约账户
    
    - Dapp可以通过Earth的API接口进行转账操作

  version: 0.0.2
  title: Dapp API 文档
...
paths:
...
  /auth/login:
    post:
      summary: 使用用户名密码登陆，返回JWT token
      description: 登陆
      tags:
        - Auth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        description: 用户名密码
        required: true
  /auth/verify:
    post:
      summary: 验证token是否有效
      description: 验证某个token是否有效，如果有效，返回这个token的数据，如果无效，返回401
      tags:
        - Auth
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  securitySchemes:
    Bearer:
      type: 'apiKey'
      name: Authorization
      in: header
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: string
          description: ''
        payload:
          type: object
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: 用户名
        password:
          type: string
          description: 密码，目前版本密码格式不加validation
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - payload
          properties:
            payload:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: 登陆成功返回的JWT token
...